Separate out mtlx converter so can be called independent of geometry loading as a separate utilitiy.
